// packages/ghs-core/src/vendor/game/model/data/AttackModifier.ts
export enum AttackModifierType {
  plus = 'plus',
  plus0 = 'plus0',
  plus1 = 'plus1',
  plus2 = 'plus2',
  plus3 = 'plus3',
  plus4 = 'plus4',
  plusX = 'plusX',
  minus = 'minus',
  minus1 = 'minus1',
  minus2 = 'minus2',
  null = 'null',
  double = 'double',
  bless = 'bless',
  curse = 'curse',
  minus1extra = 'minus1extra',
  empower = 'empower',
  enfeeble = 'enfeeble',
  invalid = 'invalid',
  townguard = 'townguard',
  wreck = 'wreck',
  success = 'success',
  imbue = 'imbue',
  advancedImbue = 'advancedImbue'
}

export enum AttackModifierValueType {
  default = 'default',
  plus = 'plus',
  minus = 'minus',
  multiply = 'multiply'
}

export class AttackModifier {
  id: string;
  type: AttackModifierType;
  value: number = 0;
  valueType: AttackModifierValueType;
  shuffle: boolean = false;
  effects: AttackModifierEffect[];
  rolling: boolean;
  active: boolean;
  revealed: boolean = false;
  character: boolean = false;

  constructor(
    type: AttackModifierType,
    value: number = 0,
    valueType: AttackModifierValueType = AttackModifierValueType.plus,
    id: string | undefined = undefined,
    effects: AttackModifierEffect[] = [],
    rolling: boolean = false,
    active: boolean = false,
    shuffle: boolean | undefined = undefined
  ) {
    this.type = type;
    this.value = value;
    this.valueType = valueType;
    if (id) {
      this.id = id;
    } else if (type === AttackModifierType.townguard) {
      this.id = 'tg-' + valueType + value;
    } else if (type === AttackModifierType.wreck || type === AttackModifierType.success) {
      this.id = 'tg-' + type;
    } else {
      this.id = type !== AttackModifierType.plus && type !== AttackModifierType.minus ? type : type + value;
    }
    this.effects = effects;
    this.rolling = rolling;
    this.active = active;
    this.shuffle = shuffle || false;
    switch (type) {
      case AttackModifierType.plus0:
        this.value = 0;
        break;
      case AttackModifierType.plus1:
        this.value = 1;
        break;
      case AttackModifierType.plus2:
        this.value = 2;
        break;
      case AttackModifierType.plus3:
        this.value = 3;
        break;
      case AttackModifierType.plus4:
        this.value = 4;
        break;
      case AttackModifierType.minus:
        this.valueType = AttackModifierValueType.minus;
        break;
      case AttackModifierType.minus1:
        this.valueType = AttackModifierValueType.minus;
        this.value = 1;
        break;
      case AttackModifierType.minus2:
        this.valueType = AttackModifierValueType.minus;
        this.value = 2;
        break;
      case AttackModifierType.null:
        this.valueType = AttackModifierValueType.multiply;
        this.value = 0;
        if (shuffle === undefined) {
          this.shuffle = true;
        }
        break;
      case AttackModifierType.double:
        this.valueType = AttackModifierValueType.multiply;
        this.value = 2;
        if (shuffle === undefined) {
          this.shuffle = true;
        }
        break;
      case AttackModifierType.bless:
        this.valueType = AttackModifierValueType.multiply;
        this.value = 2;
        break;
      case AttackModifierType.curse:
        this.valueType = AttackModifierValueType.multiply;
        this.value = 0;
        break;
      case AttackModifierType.minus1extra:
        this.valueType = AttackModifierValueType.minus;
        this.value = 1;
        break;
      case AttackModifierType.wreck:
        this.valueType = AttackModifierValueType.minus;
        this.value = 0;
        break;
      case AttackModifierType.success:
        this.value = 0;
        break;
    }
  }
}
// ...
export class AttackModifierEffect {
  type: AttackModifierEffectType;
  value: string;
  hint: string;
  effects: AttackModifierEffect[];
  icon: boolean = false;

  constructor(
    type: AttackModifierEffectType,
    value: string = '',
    hint: string = '',
    effects: AttackModifierEffect[] = [],
    icon: boolean = false
  ) {
    this.type = type;
    this.value = value;
    this.hint = hint;
    this.effects = effects;
    this.icon = icon;
  }
}
// ...
export const defaultAttackModifier: AttackModifier[] = [
  new AttackModifier(AttackModifierType.plus0),
  new AttackModifier(AttackModifierType.plus1),
  new AttackModifier(AttackModifierType.minus1),
  new AttackModifier(AttackModifierType.plus2),
  new AttackModifier(AttackModifierType.minus2),
  new AttackModifier(AttackModifierType.double),
  new AttackModifier(AttackModifierType.null),
  new AttackModifier(AttackModifierType.bless),
  new AttackModifier(AttackModifierType.curse),
  new AttackModifier(AttackModifierType.minus1extra)
];

export const defaultAttackModifierCards: AttackModifier[] = [
  // 6x +0
  new AttackModifier(AttackModifierType.plus0),
  new AttackModifier(AttackModifierType.plus0),
  new AttackModifier(AttackModifierType.plus0),
  new AttackModifier(AttackModifierType.plus0),
  new AttackModifier(AttackModifierType.plus0),
  new AttackModifier(AttackModifierType.plus0),
  // 5x +1
  new AttackModifier(AttackModifierType.plus1),
  new AttackModifier(AttackModifierType.plus1),
  new AttackModifier(AttackModifierType.plus1),
  new AttackModifier(AttackModifierType.plus1),
  new AttackModifier(AttackModifierType.plus1),
  // 5x -1
  new AttackModifier(AttackModifierType.minus1),
  new AttackModifier(AttackModifierType.minus1),
  new AttackModifier(AttackModifierType.minus1),
  new AttackModifier(AttackModifierType.minus1),
  new AttackModifier(AttackModifierType.minus1),
  // 1x +2
  new AttackModifier(AttackModifierType.plus2),
  // 1x -2
  new AttackModifier(AttackModifierType.minus2),
  // 1x x2
  new AttackModifier(AttackModifierType.double),
  // 1x x0
  new AttackModifier(AttackModifierType.null)
];
// ...
export class AttackModifierDeck {
  attackModifiers: AttackModifier[];
  current: number;
  cards: AttackModifier[];
  discarded: number[] = [];
  active: boolean = true;
  lastVisible: number = 0;
  state: 'advantage' | 'disadvantage' | undefined;
  bb: boolean;

  constructor(attackModifiers: AttackModifier[] | undefined = undefined, bb: boolean = false) {
    this.bb = bb;
    this.attackModifiers = (
      attackModifiers ? attackModifiers.filter((am, index, self) => this.bb || self.indexOf(am) === index) : defaultAttackModifier
    ).map((am) => Object.assign(new AttackModifier(am.type), am));
    this.current = -1;
    this.cards = (attackModifiers || defaultAttackModifierCards).map((am) => Object.assign(new AttackModifier(am.type), am));
  }

  toModel(): GameAttackModifierDeckModel {
    return new GameAttackModifierDeckModel(
      this.current,
      this.cards.map((attackModifier) => attackModifier && attackModifier.id),
      this.discarded,
      this.cards.map((attackModifier) => attackModifier && attackModifier.revealed),
      this.active,
      this.lastVisible,
      this.state,
      this.bb
    );
  }

  merge(attackModifierDeck: AttackModifierDeck) {
    this.attackModifiers = attackModifierDeck.attackModifiers;
    this.current = attackModifierDeck.current;
    this.cards = attackModifierDeck.cards;
    this.discarded = attackModifierDeck.discarded;
    this.lastVisible = attackModifierDeck.lastVisible;
    this.state = attackModifierDeck.state;
    this.bb = attackModifierDeck.bb;
  }
}

export class GameAttackModifierDeckModel {
  current: number;
  cards: string[];
  discarded: number[];
  revealed: boolean[];
  active: boolean;
  lastVisible: number;
  state: 'advantage' | 'disadvantage' | undefined;
  bb: boolean;
  // migration
  disgarded: number[];

  constructor(
    current: number,
    cards: string[],
    discarded: number[],
    revealed: boolean[],
    active: boolean,
    lastVisible: number = 0,
    state: 'advantage' | 'disadvantage' | undefined = undefined,
    bb: boolean = false
  ) {
    this.current = current;
    this.cards = cards;
    this.discarded = JSON.parse(JSON.stringify(discarded));
    this.revealed = JSON.parse(JSON.stringify(revealed));
    this.active = active;
    this.lastVisible = lastVisible;
    this.state = state;
    this.bb = bb;
    // migration
    this.disgarded = JSON.parse(JSON.stringify(discarded));
  }
}