// packages/ghs-core/src/vendor/game/model/Game.ts
export enum GameState {
  draw = 'draw',
  next = 'next'
}